backend/requirements.txt
fastapi[standard]
sqlmodel==0.0.24
uvicorn
passlib[bcrypt]
bcrypt==4.1.2
python-dotenv==1.1.0
python-jose
python-multipart
Faker==37.1.0
sqlalchemy
asyncpg
psycopg2-binary

